+++ dev/entities.ts
@@ -1,7 +1,8 @@
 import { InferSelectModel } from "drizzle-orm";
-import { addressTable, categoryTable, postTable, userTable } from "./drizzle/schema.js";
+import { addressTable, categoryTable, postTable, tagTable, userTable } from "./drizzle/schema.js";
 
 export type Category = InferSelectModel<typeof categoryTable>;
-export type Post = InferSelectModel<typeof postTable> & { category: Category | null };
+export type Tag = InferSelectModel<typeof tagTable>;
+export type Post = InferSelectModel<typeof postTable> & { category: Category | null; tags: Tag[]; user: User };
 export type Address = InferSelectModel<typeof addressTable> & { user: User };
 export type User = InferSelectModel<typeof userTable> & { posts: Post[]; address: Address | null };